feat: implement data cleaning
We need to clean the data before we can do any analysis.  All the values
in each column must be greater than zero.  Therefore, any row that
contains values less than 1 will be removed.  Rows with null values will
be removed as well.  The statistics above shows that the minimum of Fin
SF is 0.  Finally, we make sure that # Rooms is at least equal to the
sum of # of Bedrooms and # Baths.